requirements.txt
opencv-python==4.5.3.56
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
scipy==1.7.1
tensorflow==2.6.0  # Added TensorFlow for Keras
Pillow==8.2.0
